% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circGsea.R
\name{circGsea}
\alias{circGsea}
\title{Circular Gene Set Enrichment Analysis Plot}
\usage{
circGsea(
  object = NULL,
  geneSetID = NULL,
  htCol = c("#08519C", "#A50F15"),
  segmentsCol = c("#0099CC", "#FF0033"),
  bgCol = "grey80",
  curveCol = "#CC3333",
  pointSize = 0.5,
  addPoints = FALSE,
  type = c("c", "h", "m"),
  addGeneRank = FALSE,
  markGene = NULL,
  markGeneSize = 0.5,
  GoIDfacing = "inside",
  GoIDsize = 0.75,
  addDescription = FALSE,
  descriptionFacing = "bending.outside",
  descripShift = 1,
  descripGap = 0.3,
  descripLength = 40,
  descripSize = 0.75
)
}
\arguments{
\item{object}{An object containing enrichment results from clusterProfiler.}

\item{geneSetID}{A vector of gene set IDs to be plotted.}

\item{htCol}{A vector of two colors for the heatmap gradient. Default is c("#08519C", "#A50F15").}

\item{segmentsCol}{A vector of two colors for the segment gradient. Default is c("#0099CC", "#FF0033").}

\item{bgCol}{Background color for the sectors. Default is "grey80".}

\item{curveCol}{Color for the runningScore curve. Default is "#CC3333".}

\item{pointSize}{Size of the points on the runningScore curve. Default is 0.5.}

\item{addPoints}{Logical indicating whether to add points to the runningScore curve. Default is FALSE.}

\item{type}{Type of plot to display: 'c' for heatmap only, 'h' for segments and heatmap, 'm' for segments only. Default is "c".}

\item{addGeneRank}{Logical indicating whether to add a track for gene ranks. Default is FALSE.}

\item{markGene}{A vector of gene symbols to be marked on the plot. Default is NULL.}

\item{markGeneSize}{Size of the marked gene symbols. Default is 0.5.}

\item{GoIDfacing}{Facing direction of the gene set ID labels. Default is "inside".
Additional choice are "inside", "outside", "reverse.clockwise", "clockwise","downward", "bending",
"bending.inside", "bending.outside".}

\item{GoIDsize}{Size of the gene set ID labels. Default is 0.75.}

\item{addDescription}{Logical indicating whether to add term descriptions. Default is FALSE.}

\item{descriptionFacing}{Facing direction of the term description text. Default is "bending.outside".
Additional choice are "inside", "outside", "reverse.clockwise", "clockwise","downward", "bending",
"bending.inside", "bending.outside".}

\item{descripShift}{Shift of the term description text. Default is 1.}

\item{descripGap}{Gap between lines of the term description text. Default is 0.3.}

\item{descripLength}{Maximum line length for the term description text. Default is 40.}

\item{descripSize}{Size of the term description text. Default is 0.75.}
}
\value{
A circular gene set enrichment analysis plot.
}
\description{
This function creates a circular plot for gene set enrichment analysis.
}
