% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2gseaResult.R
\name{dfGO2gseaResult}
\alias{dfGO2gseaResult}
\title{Create gseaResult Object from Enrichment Analysis Output Dataframe}
\usage{
dfGO2gseaResult(
  enrich.df = NULL,
  geneList = NULL,
  OrgDb = NULL,
  keytype = "ENTREZID",
  setType = c("BP", "CC", "MF", "ALL"),
  pvalueCutoff = 0.05,
  eps = 1e-10,
  pAdjustMethod = "BH",
  exponent = 1,
  minGSSize = 120,
  maxGSSize = 500
)
}
\arguments{
\item{enrich.df}{A data frame containing the results of enrichment analysis.}

\item{geneList}{A decreasing sorted vector of gene list.}

\item{OrgDb}{The organism-specific annotation database.}

\item{keytype}{The type of gene identifier used in the analysis (default is "ENTREZID").}

\item{setType}{The type of GO ontology to use ("BP" for Biological Process, "CC" for Cellular Component,
"MF" for Molecular Function, or "ALL" for all ontologies).}

\item{pvalueCutoff}{The p-value cutoff for significance (default is 0.05).}

\item{eps}{A small value to avoid division by zero (default is 1e-10).}

\item{pAdjustMethod}{The p-value adjustment method (default is "BH" for Benjamini-Hochberg).}

\item{exponent}{The exponent for weighting p-values (default is 1).}

\item{minGSSize}{The minimum gene set size (default is 120).}

\item{maxGSSize}{The maximum gene set size (default is 500).}
}
\value{
A gseaResult object containing the results of GSEA.
}
\description{
This function creates a gseaResult object from the results of enrichment analysis
using gene set enrichment analysis (GSEA) for a given gene list and Gene Ontology (GO) data.
}
