% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2gseaResult.R
\name{dfKEGG2gseaResult}
\alias{dfKEGG2gseaResult}
\title{Create gseaResult Object from KEGG Enrichment Analysis Output Dataframe}
\usage{
dfKEGG2gseaResult(
  enrich.df = NULL,
  geneList = NULL,
  organism = "hsa",
  keytype = "kegg",
  setType = "KEGG",
  use_internal_data = FALSE,
  pvalueCutoff = 0.05,
  eps = 1e-10,
  pAdjustMethod = "BH",
  exponent = 1,
  minGSSize = 10,
  maxGSSize = 500
)
}
\arguments{
\item{enrich.df}{A data frame containing the results of KEGG enrichment analysis.}

\item{geneList}{A decreasing sorted vector of gene identifiers.}

\item{organism}{The organism for KEGG enrichment analysis (default is "hsa" for Homo sapiens).}

\item{keytype}{The type of gene identifier used in the analysis (default is "kegg").}

\item{setType}{The type of enrichment analysis ("KEGG" for KEGG pathways).}

\item{use_internal_data}{Logical value indicating whether to use internal KEGG data (default is FALSE).}

\item{pvalueCutoff}{The p-value cutoff for significance (default is 0.05).}

\item{eps}{A small value to avoid division by zero (default is 1e-10).}

\item{pAdjustMethod}{The p-value adjustment method (default is "BH" for Benjamini-Hochberg).}

\item{exponent}{The exponent for weighting p-values (default is 1).}

\item{minGSSize}{The minimum gene set size (default is 10).}

\item{maxGSSize}{The maximum gene set size (default is 500).}
}
\value{
A gseaResult object containing the results of KEGG enrichment analysis.
}
\description{
This function creates a gseaResult object from the results of KEGG enrichment analysis
for a given gene list and organism.
}
