% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotGsea.R
\name{dotplotGsea}
\alias{dotplotGsea}
\title{dotplotGsea}
\arguments{
\item{data}{GSEA enrich object from clusterProfiler, defalut is NULL.}

\item{pval}{pvalue cutoff to select significant terms, defalut is NULL.}

\item{pajust}{adjusted pvalue cutoff to select significant terms, defalut is 0.05.}

\item{order.by}{the X axis, defalut is "GeneRatio".}

\item{str.width}{the width of term name, defalut is 50.}

\item{base_size}{theme base size, defalut is 12.}

\item{topn}{show the top terms, defalut is NULL.}

\item{scales}{facet scales, defalut is "free_x".}

\item{add.seg}{whether add segment line to point, defalut is "FALSE".}

\item{line.col}{segment line color, defalut is "grey80".}

\item{line.size}{segment line size, defalut is 1.5.}

\item{line.type}{segment line type, defalut is "solid".}
}
\value{
a ggplot object.
}
\description{
dotplotGsea
}
\author{
Jun Zhang
}
