% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_ridge_plot.R
\name{enrich_ridge_plot}
\alias{enrich_ridge_plot}
\title{Generate Enrichment Ridge Plot}
\usage{
enrich_ridge_plot(
  object = NULL,
  terms_ID = NULL,
  gene_list = NULL,
  term_name_width = 60,
  geom_density_ridges_params = list()
)
}
\arguments{
\item{object}{An \code{enrichResult} or \code{gseaResult} object obtained from enrichment analysis.}

\item{terms_ID}{A vector of term IDs (corresponding to enriched pathways or terms) to be included in the plot.}

\item{gene_list}{A data frame with two columns("id","logfc") providing the gene fold changes (log fold changes).}

\item{term_name_width}{(optional) Maximum width of the term name for wrapping text in the y-axis. Default is 60.}

\item{geom_density_ridges_params}{(optional) A list of additional parameters passed to \code{geom_density_ridges} to customize the ridge plot appearance.}
}
\value{
A \code{ggplot} object which contains the ridge plot visualizing the log fold change distribution of enriched terms.
}
\description{
This function creates a ridge plot for visualizing gene enrichment analysis results using either \code{enrichResult} or \code{gseaResult} objects.
Ridge plots allow for the visualization of the distribution of genes across multiple terms and enriched pathways.
}
\details{
This function accepts two types of enrichment result objects: \code{enrichResult} and \code{gseaResult}. For both types, the function handles gene identifiers
(either raw gene IDs or human-readable names) and generates the appropriate data frames with corresponding log fold change values.
It then filters and plots the specified terms.

The ridge plot helps visualize how genes for each term distribute across log fold change values, and it also highlights the statistical
significance of the terms using a color gradient based on the p-value.
}
