% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAmultiGP.R
\name{GSEAmultiGP}
\alias{GSEAmultiGP}
\title{Perform GSEA Multi-group Plotting}
\usage{
GSEAmultiGP(
  gsea_list = NULL,
  geneSetID = NULL,
  exp_name = NULL,
  addPval = FALSE,
  curve.col = NULL,
  curve.linewidth = 1,
  kegg = FALSE,
  lineSize = 1,
  base_size = 12,
  nesDigit = 2,
  pDigit = 2,
  pvalX = 0.9,
  pvalY = 0.9,
  rect.bm.col = c("#CC3333", "white", "#003366"),
  subplot.heights = c(0.4, 0.2, 0.08),
  legend.position = c(0.85, 0.85),
  rect.bm.label = c("Up regulated", "Down regulated"),
  breaks.n = 6
)
}
\arguments{
\item{gsea_list}{A list of GSEA results for multiple experiments.}

\item{geneSetID}{The ID of the gene set to be visualized.}

\item{exp_name}{Names of the experiments corresponding to the GSEA results.}

\item{addPval}{Logical, indicating whether to add NES (Normalized Enrichment Score)
and p-value labels to the plot. Default is FALSE.}

\item{curve.col}{A vector of colors for the curves representing different
experiments. If NULL, random colors are assigned.}

\item{curve.linewidth}{The curve linewidth. Default is 1.}

\item{kegg}{Logical, indicating whether the gene set is a KEGG pathway.
Default is FALSE.}

\item{lineSize}{The size of the lines in the enrichment score curve plot.
Default is 1.}

\item{base_size}{The base font size for the plot. Default is 12.}

\item{nesDigit}{The number of digits to round NES values to. Default is 2.}

\item{pDigit}{The number of digits to round p-values to. Default is 2.}

\item{pvalX}{The X-coordinate for placing p-value labels on the plot. Default is 0.9.}

\item{pvalY}{The Y-coordinate for placing p-value labels on the plot. Default is 0.9.}

\item{rect.bm.col}{A vector of colors for the bottom rectangle representing
up-regulated and down-regulated genes. Default colors c("#CC3333","white","#003366")
are provided.}

\item{subplot.heights}{Heights of subplots in the multi-group plot. Default
values c(0.4,0.2,0.08) are provided.}

\item{legend.position}{The position of the legend in the plot. Default is (0.85, 0.85).}

\item{rect.bm.label}{Labels for the bottom rectangle, specifying "Up regulated"
and "Down regulated". Default labels are provided.}

\item{breaks.n}{The number of X axis breaks. Default is 6.}
}
\value{
Returns a multi-panel plot for GSEA results.
}
\description{
This function generates a multi-group plot for Gene Set Enrichment Analysis
(GSEA) results.
}
\examples{
\dontrun{
# Example Usage
result <- GSEAmultiGP(gsea_list = gsea_results,
                      geneSetID = "gene_set_1",
                      exp_name = c("Exp1", "Exp2"),
                      addPval = TRUE,
                      curve.col = c("red", "blue"),
                      pvalX = 0.9,
                      pvalY = 0.9,
                      rect.bm.col = c("#CC3333", "white", "#003366"),
                      subplot.heights = c(0.4, 0.2, 0.08),
                      legend.position = c(0.85, 0.85),
                      rect.bm.label = c("Up regulated", "Down regulated"))
                      }

}
\author{
Jun Zhang
}
