% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotGsea.R, R/gseaNb.R
\name{gseaNb}
\alias{gseaNb}
\title{gseaNb}
\arguments{
\item{data}{GSEA enrich object from clusterProfiler, defalut is NULL.}

\item{pval}{pvalue cutoff to select significant terms, defalut is NULL.}

\item{pajust}{adjusted pvalue cutoff to select significant terms, defalut is 0.05.}

\item{order.by}{the X axis, defalut is "GeneRatio".}

\item{str.width}{the width of term name, defalut is 50.}

\item{base_size}{theme base size, defalut is 12.}

\item{topn}{show the top terms, defalut is NULL.}

\item{scales}{facet scales, defalut is "free_x".}

\item{add.seg}{whether add segment line to point, defalut is "FALSE".}

\item{line.col}{segment line color, defalut is "grey80".}

\item{line.size}{segment line size, defalut is 1.5.}

\item{line.type}{segment line type, defalut is "solid".}

\item{object}{GSEA enrich results.}

\item{subPlot}{which plot to show, 1/2/3, default is 3.}

\item{lineSize}{curve line size. default is 0.8.}

\item{geneSetID}{which pathway name to plot.}

\item{rmSegment}{whether to remove segment on the curve plot, default is FALSE.}

\item{termWidth}{the width or the term name, defalut is 40.}

\item{segCol}{segment color on the curves, defalut is "red".}

\item{addGene}{whether add gene name on the curve, defalut is FALSE.}

\item{geneCol}{gene name label color, defalut is NULL.}

\item{arrowAngle}{arrow angle, defalut is 20.}

\item{arrowLength}{arrow line length, defalut is 0.2.}

\item{arrowEnd}{arrow end, defalut is "last".}

\item{arrowType}{arrow type, defalut is "closed".}

\item{curveCol}{curve color, defalut is c("#76BA99", "#EB4747").}

\item{htCol}{heatmap color, defalut is c("#08519C", "#A50F15").}

\item{rankCol}{gene rank fill color, defalut is c("#08519C", "white", "#A50F15").}

\item{rankSeq}{gene rank plot X axis breaks, defalt is 5000.}

\item{htHeight}{the relative height when "subplot = 2" to the vertical line plot, defalut is 0.3.}

\item{force}{the gene label force, refer to geom_text_repel function, defalut is 20.}

\item{max.overlaps}{refer to geom_text_repel function, defalut is 50.}

\item{geneSize}{gene label text size, defalut is 4.}

\item{newGsea}{whether show new style of plot, defalut is FALSE.}

\item{addPoint}{new style plot with point layer, defalut is TRUE.}

\item{newCurveCol}{new style plot curve color, defalut is c("#336699", "white", "#993399").}

\item{newHtCol}{new style plot heatmap color, defalut is c("#336699", "white", "#993399").}

\item{rmHt}{whether remove new style plot heatmap, defalut is FALSE.}

\item{addPval}{whether add pvalue and NES, defalut is FALSE.}

\item{pvalX}{set pvalue label x position, defalut is 0.9.}

\item{pvalY}{set pvalue label y position, defalut is 0.9.}

\item{pvalSize}{set pvalue label text size, defalut is 4.}

\item{pCol}{pvalue label color, defalut is "grey30".}

\item{pHjust}{pvalue label hjust, defalut is 1.}

\item{rmPrefix}{whether remove GO term prefix like "GOBP/KEGG/CC/MF_*", defalut is TRUE.}

\item{nesDigit}{the NES score digits retained, defalut is 2.}

\item{pDigit}{the pvalue and pajust value digits retained, defalut is 2.}

\item{markTopgene}{whether add top n genes on plot, defalut is FALSE.}

\item{topGeneN}{the number of genes to be marked on plot, defalut is 5.}

\item{kegg}{whether input is gseKEGG object, defalut is FALSE.}
}
\value{
a ggplot object.

ggplot2 object
}
\description{
gseaNb

gseaNb
}
\examples{
\dontrun{dotplotGsea(data = ego3,topn = 10)}
\donttest{# load data
test_data <- system.file("extdata", "gseaRes.RDS", package = "GseaVis")
gseaRes <- readRDS(test_data)

# all plot
gseaNb(object = gseaRes,
      geneSetID = 'GOBP_NUCLEOSIDE_DIPHOSPHATE_METABOLIC_PROCESS')
}
}
\author{
Jun Zhang
}
