% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsInfo.R
\name{gsInfo}
\alias{gsInfo}
\title{gsInfo}
\usage{
gsInfo(object, geneSetID)
}
\arguments{
\item{object}{gseaResult object}

\item{geneSetID}{gene set ID}
}
\value{
data.frame
}
\description{
gsInfo
}
\author{
Guangchuang Yu
}
