% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsInfoNew.R
\name{gsInfoNew}
\alias{gsInfoNew}
\title{gsInfoNew}
\usage{
gsInfoNew(geneList = NULL, geneSetID = NULL, geneSet = NULL, exponent = 1)
}
\arguments{
\item{geneList}{geneList for GSEA software outputs which is saved in
"gseaOutputs/ranked_gene_list_treat_versus_control*.tsv".}

\item{geneSetID}{gene set ID.}

\item{geneSet}{enrichment term sets for GSEA software outputs
which is saved in "*/edb/gene_sets".}

\item{exponent}{weight of each step, defalut is 1.}
}
\value{
a data.frame
}
\description{
gsInfoNew
}
\author{
Guangchuang Yu, modified by JunZhang
}
