% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaNb.R
\docType{data}
\name{intergrated}
\alias{intergrated}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{list} of length 3.
}
\usage{
intergrated
}
\description{
This is a test data for this package
test data describtion
}
\author{
JunZhang
}
\keyword{datasets}
