% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGseaFile.R
\name{readGseaFile}
\alias{readGseaFile}
\title{readGseaFile}
\arguments{
\item{filePath}{the path of the GSEA software enrichment outputs, defalut is NULL.}
}
\value{
a list contains meta(intergated enrichment results),
glist(the ordered gene lists), gset(all background enrichment terms).
}
\description{
readGseaFile
}
