% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoGsea.R
\name{volcanoGsea}
\alias{volcanoGsea}
\title{volcanoGsea}
\usage{
volcanoGsea(
  data = NULL,
  NES.cutoff = 1,
  pvalue.cutoff = NULL,
  p.adjust.CUTOFF = 0.05,
  nudge.y = c(0, 0),
  topN = 5,
  point.size = 3,
  point.color = c("#CC3333", "#CCCCCC", "#0099CC"),
  ...
)
}
\arguments{
\item{data}{GSEA enrich object from clusterProfiler, defalut is NULL.}

\item{NES.cutoff}{NES cutoff to select significant terms, defalut is 1.}

\item{pvalue.cutoff}{pvalue cutoff to select significant terms, defalut is NULL.}

\item{p.adjust.CUTOFF}{adjusted pvalue cutoff to select significant terms, defalut is 0.05.}

\item{nudge.y}{y shift to ajust label, defalut is c(0,0).}

\item{topN}{top term to show, defalut is 5.}

\item{point.size}{point size, defalut is 3.}

\item{point.color}{point color, defalut is c('#CC3333','#CCCCCC','#0099CC').}

\item{...}{other arguments passed by geom_text_repel.}
}
\value{
a ggplot object.
}
\description{
volcanoGsea
}
\author{
Jun Zhang
}
