% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyGoPlot.R
\name{sankeyGoPlot}
\alias{sankeyGoPlot}
\title{sankeyGoPlot}
\usage{
sankeyGoPlot(
  goData = NULL,
  topGenes = 5,
  sankeyExpand = c(0.5, 1),
  nodeSize = 2.5,
  nodeColor = NULL,
  goCol = NULL,
  xShift = 0.05,
  downShift = 4.5,
  upShift = 0.25
)
}
\arguments{
\item{goData}{GO data frame from Clusterprofiler.}

\item{topGenes}{the top genes to be shown for each term, default 5.}

\item{sankeyExpand}{the sankey plot expand for left and side, default c(0.5,1).}

\item{nodeSize}{node text size for sankey plot, default 2.5.}

\item{nodeColor}{node fill color for sankey plot, default NULL.}

\item{goCol}{go color for go plot, default NULL.}

\item{xShift}{the shift on horizontal for go plot, default 0.05.}

\item{downShift}{the shift on vertical for go plot, default 4.5.}

\item{upShift}{the shift on vertical for go plot, default 0.25.}
}
\value{
ggplot obeject
}
\description{
sankeyGoPlot
}
